Simplified dRAID map allocations
The idea here is that by more carefully constructing the raidz maps
for dRAID we can largely avoid the need to modify the generic raidz
parity calculation code.  To do this we construct one of three
different mappings as appropriate.

  1) Write mapping.  Always a full stripe width and includes
     zero filled skip sectors which are written to disk.

  2) Read mapping.  Only includes columns with data and not skip
     sectors which can be ignored for a normal read.

  3) Scrub mapping.  Allways a full stripe width and includes
     buffer space to read in and verify the skip sectors.

This change makes use of the new gang ABDs to effeciently construct
the full column as a single ABD.  This uncovered a few issues in the
gang ABD handling which were addressed.  The I/O aggregation logic
was updated to not aggregate gang ABDs because nesting gang ABDs is
not currently supported.  These IOs are rare and it does not effect
performance.

Other changes:

* Renamed vdev_draid_get_permutation() to vdev_draid_permute_id()
  and simplifed to remove the need for a memory allocation.

* Moved vdev_draid_debug_zio() function and renamed.  This can be
  retired entirely once dRAID has been fully debugged.

* Removed the need for the vdc->vdc_zero_abd.  The generic zero
  ABD is used instead.

* Removed custom skip sector functions for dRAID.  The generic
  raidz logic is now relied upon exclusively for repair.  e.g.
  vdev_draid_restore_skip_sectors, vdev_draid_restore_skip_sectors,
  vdev_draid_fix_skip_sectors.

* Reverted raidz changes from previous commits.

* Reenable mmp testing is ztest, disabled in original dRAID commit.

* vdev_raidz_reconstruct_general() incorrectly assumed that it
  was sufficient to check the first column of data to determine
  if it was linear.  Update the check to verify al data columns.

* updated redundancy_draid_spare.ksh test to use a traditional
  spare for now until the issue with the distribute spare can
  be determined.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>