Add gang ABD child to parent gang ABD
By design a gang ABD can not have another gang ABD as a child. This is
to make sure the logical offset in a gang ABD is consistent with the
individual ABDS it contains as children. If a gang ABD is added as a
child of a gang ABD we will add the individual children of the gang ABD
to the parent gang ABD. Ths allows for a consistent view of offsets
within the parent gang ABD.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>