Fix rebuild with multiple distributed spares
If this child is a distributed spare and we're resilvering
then this offset might reside on the vdev being replaced.
In which case this data must be written to the new device.
Failure to do so would result in checksum errors when the
old device is detached and the pool is scrubbed.

Furthermore, when performing a healing resilver these DVAs would
have been excluded by the DTL check in vdev_dtl_need_resilver().
To ensure they are resilvered for dRAID this check was moved in
to vdev->vdev_op_need_resilver callback to allow each vdev type
to perform the correct check.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>