Updated math and consolidate derived constants
Updated math for the functions vdev_draid_group_degraded() and
vdev_draid_spare_get_child.  We're still seeing checksum errors
and stack overruns so this isn't a complete fix but it is a step
in the right direction.

Provide the following constants as vdc variables.  This has the
advantage of forcing consistent terminology, simplifying several
functions, and the values can be sanity checked when initialized.

    uint64_t vdc_groupwidth;        /* = vdc_data + vdc_parity */
    uint64_t vdc_ndisks;            /* = vdc_children - vdc_spares */
    uint64_t vdc_groupsz;           /* = vdc_groupwidth * DRAID_SLICESIZE */
    uint64_t vdc_slicesz;           /* = vdc_groupsz * vdc_groups */

Use physical_offset as the argument name for functions which expect
the physical offset to avoid any confusion.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>