Remove offset from vdev_psize_to_asize()
Now that all dRAID groups are the same size there is no longer a need
to pass the offset to determine a specific group id and group size.
Remove the offset from the prototype and revert all callers.  This is
particularly nice since the common code no longer need to deal with
this specific need of dRAID.  It is more like the existing vdev types.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>