Add handling for spa_config_parse() error
A small format change in the name of a dRAID configuration nvlist
key made during development process resulted in a now invalid MOS
config for pools created with the earlier code.  This resulted
an in error when trying to parse it during import and the VERIFY
was hit.  While this invalid config was direct consequence on the
dRAID format change it's not inconceivable for MOS config to be
incorrect in some other way.  In which case it's best the handle
the error and fail the import cleanly rather than panic the system.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>