Use .vdev_op_need_resilver callback
Further update the existing .vdev_op_need_resilver callback such
that it can be used generically for both healing and sequential
resilvers.

Updated the dRAID code with an implementaiton which should work
for either case.  However, additional verification of the DTL
handling for dRAID is needed.

Added back logic to always resilver when multiple distributed
spares are active.  For the moment this is still needed.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>