Replace zfs_rebuild_queue_limit
Using a basic queue limit worked well enough when sequential resilvers
only supported mirror devices.  However, in order to accomidate dRAID
where the repair IO size is directly related to the group width it's
preferable to use an inflight byte limit.  We apply the same heuristic
used in the dsl_scan code and a limit of 4MB per child vdev.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>