Tweak defaults for map generation
Testing has shown that in practice a single pass in sufficent to
generate a good map.  The previous default of 3 passes was overly
pessimistic, decrease it to 1 to speed up zpool create.  This is
particularly helpful when there are large number of children.

More important is having enough permuations for the layout.
Increase the default to 64 for pools with 100 children or less.
Even with more permutations a 128 child pool's label is only
40% full.  However, we leave it at 32 for larger pools to avoid
consuming all the label space.