Restore most of the original allocator code
When using the tiling dRAID layout we can always calculate the asize.
This means we don't need to extend the metaslab code to handle a
different asize based on the offset.  All we need to do is ensure
the starting offset is properly aligned and that the block never
spans groups.  Since we're allocating from a metaslab we're already
guarenteed the allocation will not span metaslabs.  Lastly, ensure
the size of a metaslab is a multiple of the group size.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>